requirements.txt
python-telegram-bot==20.0
openai
httpx
python-dotenv
uvicorn
fastapi
